issue #35, #38: BasicStringBuilder/TextPool - support linux/gcc

+++ stdext/include/stlpatch.h
@@ -69,7 +69,7 @@ __forceinline _It _ConvIt(_It it)
 
 template <class _Container>
 __forceinline 
-typename _Container::value_type* resize(_Container& container, size_t newSize)
+typename _Container::value_type* resize(_Container& container, unsigned int newSize)
 	{container.resize(newSize);
 	 return _ConvIt(container.begin()); }
